[api] Add TEST_DB_DUMP env to be able to cache db dump
Useful for faster testing when writing tests locally

+++ src/api/tests/conftest.py
@@ -29,21 +29,34 @@ class TestDatabase(object):
         """
         Creates a dump of the test database into file specified in self.dump_path.
         """
+        if os.environ.get('TEST_DB_DUMP') and os.path.exists(os.environ.get('TEST_DB_DUMP')):
+            return
         with open(os.devnull, "w") as f:
             subprocess.call('createdb {uri}'.format(uri=os.environ["DB_URL"]), shell=True, stdout=f)
         DepositTestdata(db).deposit_all(test_set='integration_testing')
+
+        if os.environ.get('TEST_DB_DUMP'):
+            dump_path = os.environ['TEST_DB_DUMP']
+        else:
+            dump_path = self.dump_path
         # Note the --clean and --create flags, which will recreate db when run
-        subprocess.check_call('pg_dump {uri} --file={path} --clean --create'.format(uri=os.environ["DB_URL"], path=self.dump_path), shell=True)
-        print "Temporary database file created at {}.".format(self.dump_path)
+        subprocess.check_call('pg_dump {uri} --file={path} --clean --create'.format(uri=os.environ["DB_URL"], path=dump_path), shell=True)
+        print "Temporary database file created at {}.".format(dump_path)
 
     def refresh(self):
         """
         Wipes out whole database, and recreates a clean copy from the dump.
         """
         print "Refreshing database with data from dump"
+
+        if os.environ.get('TEST_DB_DUMP') and os.path.exists(os.environ.get('TEST_DB_DUMP')):
+            dump_path = os.environ['TEST_DB_DUMP']
+        else:
+            dump_path = self.dump_path
+
         with open(os.devnull, "w") as f:
             # Connect to template1 so we can remove whatever db name
-            subprocess.check_call('psql postgres://postgres@/template1 < {path}'.format(uri=os.environ["DB_URL"], path=self.dump_path), shell=True, stdout=f)
+            subprocess.check_call('psql postgres://postgres@/template1 < {path}'.format(uri=os.environ["DB_URL"], path=dump_path), shell=True, stdout=f)
 
     def cleanup(self):
         print "Disconnecting..."
